+++ src/submission/submission-to-clinical/stat-calculator.ts
@@ -5,7 +5,7 @@ import {
 } from '../../../src/clinical/clinical-entities';
 import { ClinicalEntitySchemaNames } from '../../common-model/entities';
 import { notEmpty, F } from '../../../src/utils';
-import { getClinicalEntitiesFromDonorBySchemaName } from './submission-to-clinical';
+import { getClinicalEntitiesFromDonorBySchemaName } from '../../common-model/functions';
 
 import { DeepReadonly } from 'deep-freeze';
 import { cloneDeep, pull } from 'lodash';
