+++ src/submission/validation-clinical/followUp.ts
@@ -7,7 +7,7 @@ import { ClinicalEntitySchemaNames } from '../../common-model/entities';
 import { DeepReadonly } from 'deep-freeze';
 import { Donor } from '../../clinical/clinical-entities';
 import _ from 'lodash';
-import { getClinicalEntitiesFromDonorBySchemaName } from '../submission-to-clinical/submission-to-clinical';
+import { getClinicalEntitiesFromDonorBySchemaName } from '../../common-model/functions';
 import { checkClinicalEntityDoesntBelongToOtherDonor } from './utils';
 
 export const validate = async (
