+++ src/submission/validation-clinical/specimen.ts
@@ -14,7 +14,7 @@ import { Donor, Specimen } from '../../clinical/clinical-entities';
 import * as utils from './utils';
 import _ from 'lodash';
 import { isEmptyString, isAbsent } from '../../utils';
-import { getSingleClinicalObjectFromDonor } from '../submission-to-clinical/submission-to-clinical';
+import { getSingleClinicalObjectFromDonor } from '../../common-model/functions';
 
 export const validate = async (
   specimenRecord: DeepReadonly<SubmittedClinicalRecord>,
