+++ src/submission/validation-clinical/treatment.ts
@@ -12,7 +12,7 @@ import { DeepReadonly } from 'deep-freeze';
 import { Donor, Treatment } from '../../clinical/clinical-entities';
 import * as utils from './utils';
 import _ from 'lodash';
-import { getSingleClinicalObjectFromDonor } from '../submission-to-clinical/submission-to-clinical';
+import { getSingleClinicalObjectFromDonor } from '../../common-model/functions';
 import { checkClinicalEntityDoesntBelongToOtherDonor } from './utils';
 
 export const validate = async (
