+++ src/submission/validation-clinical/utils.ts
@@ -19,7 +19,7 @@ import { validationErrorMessage } from '../submission-error-messages';
 import _ from 'lodash';
 import { DataRecord } from '../../lectern-client/schema-entities';
 import { Donor, ClinicalInfo } from '../../clinical/clinical-entities';
-import { getSingleClinicalEntityFromDonorBySchemanName } from '../submission-to-clinical/submission-to-clinical';
+import { getSingleClinicalEntityFromDonorBySchemanName } from '../../common-model/functions';
 import { donorDao } from '../../clinical/donor-repo';
 
 export const buildSubmissionError = (
